Changed Thread type to Long Double for more precision

+++ srcs/draw_fractals.c
@@ -6,35 +6,35 @@
 /*   By: mmondell <mmondell@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/02 09:58:16 by mmondell          #+#    #+#             */
-/*   Updated: 2021/09/15 14:44:41 by mmondell         ###   ########.fr       */
+/*   Updated: 2021/12/02 08:51:35 by mmondell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/fractol.h"
 
 /*
-*	Declares an array of 6 copy of main fractol struct
-*	Each copy if then initiated to fractol struct values (ft_memcpy)
-*	Then it is important to limit each threads to a portion of the image to 
-*	draw. i + 1 is to make sure they don't walk on each others
-*	Then you create the threads (each one will have a unique ID)
-*	Each thread will then go into the draw_pixel function and draw a portion
-*	of the image.
-*	pthread_join will make sure all threads comes back at the same 
-*	point before the next loop.
-*/
-void	create_threads(t_fractol *f)
+ *	Declares an array of 6 copy of main fractol struct
+ *	Each copy if then initiated to fractol struct values (ft_memcpy)
+ *	Then it is important to limit each threads to a portion of the image to
+ *	draw. i + 1 is to make sure they don't walk on each others
+ *	Then you create the threads (each one will have a unique ID)
+ *	Each thread will then go into the draw_pixel function and draw a portion
+ *	of the image.
+ *	pthread_join will make sure all threads comes back at the same
+ *	point before the next loop.
+ */
+void create_threads(t_fractol *f)
 {
-	t_fractol	cpy[8];
-	pthread_t	t_id[8];
-	int			i;
+	t_fractol cpy[8];
+	pthread_t t_id[8];
+	int i;
 
 	i = 0;
 	while (i < 6)
 	{
 		ft_memcpy((void *)&cpy[i], (void *)f, sizeof(t_fractol));
-		cpy[i].thread.x = ((double)(1.00 / 6)*WINW) * i;
-		cpy[i].thread.max = ((double)(1.00 / 6)*WINW) * (i + 1);
+		cpy[i].thread.x = ((long double)(1.00 / 6) * WINW) * i;
+		cpy[i].thread.max = ((long double)(1.00 / 6) * WINW) * (i + 1);
 		cpy[i].thread.y = 0;
 		i++;
 	}
@@ -52,9 +52,9 @@ void	create_threads(t_fractol *f)
 }
 
 /*
-*	Function that updates the value seen in the window dynamically
-*/
-void	fractal_updater(t_fractol *f)
+ *	Function that updates the value seen in the window dynamically
+ */
+void fractal_updater(t_fractol *f)
 {
 	if (f->fractal.max_iter <= 1)
 		f->fractal.max_iter = 1;
@@ -64,13 +64,13 @@ void	fractal_updater(t_fractol *f)
 }
 
 /*
-*	Limits the pixel draw to the min and max size of the window
-*	if the iteration equals the number of the fractal max iteration
-*	paint the pixel black. (The actual fractal set)
-*	otherwise give each RGB a different color value.
-*	RGB values have been tinkered to give a color I liked
-*/
-void	put_pixel(t_fractol *f, int x, int y, int iter)
+ *	Limits the pixel draw to the min and max size of the window
+ *	if the iteration equals the number of the fractal max iteration
+ *	paint the pixel black. (The actual fractal set)
+ *	otherwise give each RGB a different color value.
+ *	RGB values have been tinkered to give a color I liked
+ */
+void put_pixel(t_fractol *f, int x, int y, int iter)
 {
 	if (x >= 0 && x <= WINW && y >= 0 && y <= WINH)
 	{
@@ -82,27 +82,24 @@ void	put_pixel(t_fractol *f, int x, int y, int iter)
 		}
 		else
 		{
-			f->mlx.data[(x * 4) + (y * WINW * 4) + 2] = f->color.red
-				+ (iter * 2) % 256;
-			f->mlx.data[(x * 4) + (y * WINW * 4) + 1] = f->color.green
-				+ (iter * 5) % 256;
-			f->mlx.data[(x * 4) + (y * WINW * 4)] = f->color.blue
-				+ (iter * 3) % 256;
+			f->mlx.data[(x * 4) + (y * WINW * 4) + 2] = f->color.red + (iter * 2) % 256;
+			f->mlx.data[(x * 4) + (y * WINW * 4) + 1] = f->color.green + (iter * 5) % 256;
+			f->mlx.data[(x * 4) + (y * WINW * 4)] = f->color.blue + (iter * 3) % 256;
 		}
 	}
 }
 
 /*
-*	Function to draw each pixel starting at top left corner (0, 0)
-*	and giving a different color depending on the number of iterations
-*	of starting coordinates before it escapes to inifinty.
-*	That gives the fractals the beautiful colors.
-*/
-void	*pixel_draw(void *cpy)
+ *	Function to draw each pixel starting at top left corner (0, 0)
+ *	and giving a different color depending on the number of iterations
+ *	of starting coordinates before it escapes to inifinty.
+ *	That gives the fractals the beautiful colors.
+ */
+void *pixel_draw(void *cpy)
 {
-	t_fractol	*f;
-	int			temp_x;
-	int			color;
+	t_fractol *f;
+	int temp_x;
+	int color;
 
 	f = (t_fractol *)cpy;
 	f->thread.y = 0;
